+++ packages/unlock-js/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changes
 
+# 0.37.1
+
+- fixing swap and purchase
+
+# 0.37.0
+
+- adding support for actual purchase through CardPurchaser
+
+# 0.36.1
+
+- fix decimals for withdrawFromLock
+
+# 0.36.0
+
+- adding support functions for Transfer With Authorization
+- adding support for CardPurchaser contract
+
 # 0.35.5
 
 - Removed signDataPersonal